requirements.txt
GitPython
keyboard
mss
numpy
opencv-python-headless
Pillow
pygame
pywin32
tensorflow
pyautogui
customtkinter
pytesseract
